requirements.txt
# Django and Wagtail
Django>=5.0,<5.3
wagtail>=6.0,<7.0

# Database
psycopg2-binary>=2.9.5

# Image processing
Pillow>=10.0.0

# REST API
djangorestframework>=3.14.0

# Search
elasticsearch>=8.0.0

# Development tools
django-extensions>=3.2.0
django-debug-toolbar>=4.0.0

# Production
gunicorn>=21.2.0
whitenoise>=6.5.0

# Email
django-ses>=3.5.0

# Caching
redis>=4.6.0
django-redis>=5.3.0

# Monitoring
sentry-sdk>=1.32.0

# Environment
python-decouple>=3.8
